docs(select): fix FilterMultiSelect and Select docs
